// src/app/interfaces/currentRoute.interface.ts
export interface CurrentRoute {
    id:                   string;
    has_sub:              boolean;
    has_doc:              boolean;
    nombre:               string;
    nombre2:              string;
    dependencia_indice:   string;
    fonatur:              number;
    dependencia_superior: number;
  }